helper/linux: Use $(CC) instead of 'cc'
In order to compile chipsec for a non-native toolchain we need to use a
level of indirection and reference $(CC) instead of 'cc'.

By default there's no change, and things should work as before. But now
it's also possible for users to specify CC= on the command line to point
to a non-native compiler.

Signed-off-by: Matt Fleming <matt.fleming@intel.com>