Apollo Lake/Gemini Lake: Fix register definitions
This commit fixes several register/field definitions for APL and GLK. I found
these mistakes after checking against the Intel datasheets.

In some cases (such as BDSM and friends) I saw that we can just fall back to the
'common' definitions, since those match the datasheets'.

Notice that I intentionally didn't fix ABASE for GLK, so as not to conflict with
PR #890.