Convert interrupts_cmd to argparse
This commit attempts to convert the interrupts_cmd file
to use argparse (addresses #673).  Note that this commit
proposes a change to the smi command arguments by splitting
the default command into "sw" and "apmc" commands; if these
kinds of API changes are not possible this can still be
refactored to conform to the original API, albeit at the
slight cost of code cleanliness.