Workaround for function pointer complaints
* Some function pointers misuse caught by PaX RAP. These workaround fix would be useful
to the vanilla kernel since the recent TigerLake support Intel CET which might have
the same complaints if the implementation can achieve the current level as PaX RAP.

* Static call is a hack to make the function pointer writable. Please keep in mind that
this hack is only for the firmware audit reason. Be cautions if you intend to do the similar
things in production.

Thanks to Spender and PaX team points out the way to fix.

Signed-off-by: Shawn C <citypw@hardenedlinux.org>