drivers/linux: use appropriate error code on module init errors
If the symbol lookup via find_symbols() fails, init_module() returns
with -1 as an error code. This is unfortunate, however, as that value
will be passed up to userland as an error code, so should rather be a
symbolic one that can be grep'ed for (-1 would be -EPERM which doesn't
seem fitting).

Use -EOPNOTSUPP instead.

Signed-off-by: Mathias Krause <minipli@grsecurity.net>