helper/linux: don't unnecessary (un)load the kernel module
When the kernel module is already loaded, don't unload and try to load
it again.

This is not only useful during development where one might change the
kernel module code often and load it manually but also for setups that
lift the strict access restrictions to /dev/chipsec, e.g. to allow a
non-root group access to its interface.

Signed-off-by: Mathias Krause <minipli@grsecurity.net>