Introduce vendor-specific module logic
The goal is to be able to run Intel-specific modules on Intel only (and
likewise for AMD or other vendors).

The new module architecture is as follows:
- chipsec/modules/…:
  - common/…: modules loaded for all vendors and platforms
  - 8086/…: Intel-specific modules
    - common/…: modules loaded for all Intel platforms
    - ivb/…: modules specific to the IVB (Ivy Bridge) platform
    - etc.

This patch does not introduce vendor-specific HAL logic or commands.
Only modules ran by chipsec_main are in-scope.

A follow-up patch will move Intel-specific modules to
chipsec/modules/8086.

Bug: #1318
Signed-off-by: Gabriel Kerneis <gabriel.kerneis@ssi.gouv.fr>