Fix premature return at module children updating
When update_efi_tree() is called recursively to update the attributes of a module which contains children, it is possible for the children attributes to be partially updated (effectively skipped) because a premature return occurs when a UI section is encountered. This means that the children modules may have missing (i.e. None) value for (parent) GUID.

For some reason, update_efi_tree() function appears to have two purposes: 1) Update EFI tree, but also 2) Return the name (ui_string) of a given module. It should only be doing the former and separate the returning of a module's name to a different function.

Alas, it seems that the bug can be fixed by replacing the premature return statement with a "ui_string" variable assignment instead, which is then properly returned at the end of the update_efi_tree() function.