Make OsHelper.get_available_helpers return a sorted list
`OsHelper.get_available_helpers` is returning a list of helper names in
the order they are stored in the dict `self.avail_helpers`. This order
seems to come from the order in which `os.listdir(helper_dir)` reads the
content of the directory.

To make this order more predictible and meaningful, sort the list in
`get_available_helpers`. This function was only called when generating
the documentation ofi option `--helper` and this makes it more
reproducible too.

Signed-off-by: Nicolas Iooss <nicolas.iooss_git@polytechnique.org>