Fix objtool 'unannotated intra-function call'
Fixes objtool errors in stricter kernel configs

objtool likely wants the function to be annotated with
SYM_CODE_START/SYM_CODE_END (linux/linkage.h). It might be
better to rewrite the linux kernel module assembly files to use gas,
and properly annotate everything, but this should be a quick fix for
now.

Signed-off-by: J. Stephan Bentze <j.s.bentze@gmail.com>